requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bidict==0.23.1
black==24.4.2
boilerpy3==1.0.7
certifi==2024.2.2
chainlit==1.1.101
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
dataclasses-json==0.5.14
datasets==2.19.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastapi==0.110.3
fastapi-socketio==0.0.10
filelock==3.14.0
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==7.0.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.59
Lazify==0.4.0
lazy-imports==0.3.1
literalai==0.0.601
MarkupSafe==2.1.5
marshmallow==3.21.2
monotonic==1.6
more-itertools==10.2.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
posthog==3.5.0
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
syncer==2.0.3
tenacity==8.3.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
